Move quote unpickling to the core package
